Fix compile error. Move codes to  proper places.
