+++ grimbot/CmdPatterns.py
@@ -1,6 +1,36 @@
-from .CommandManager import ArgsPatternPart, CommandArgsPatternDoesntMatchException
 from typing import Sequence
 import re
+from abc import ABCMeta, abstractmethod
+
+
+class ArgsPatternPart(metaclass=ABCMeta):
+
+    @abstractmethod
+    def argsLength(self) -> int:
+        raise NotImplementedError()
+
+    @abstractmethod
+    def isMatch(self, args : Sequence[str], index : int) -> bool:
+        raise NotImplementedError()
+
+
+# コマンドの引数の長さが一致しなかった時に発生させる例外です
+class CommandLengthDoesntMatchException(Exception):
+    def __init__(self, length : int):
+        self.length = length
+
+    def __str__(self):
+        return "このコマンドは、少なくとも%d個の引数が必要です!" % self.length
+
+
+# コマンドのパターンが一致しなかった時に発生させる例外です
+class CommandArgsPatternDoesntMatchException(Exception):
+    def __init__(self, message : str):
+        self.message = message
+
+    def __str__(self):
+        return self.message
+
 
 class NumberArgsPattern(ArgsPatternPart):
 
@@ -14,6 +44,7 @@ class NumberArgsPattern(ArgsPatternPart):
             raise CommandArgsPatternDoesntMatchException("%d番目の引数が数値ではありません!" % (index + 1))
         else:
             return True
+
 
 class PositiveNumberArgsPattern(ArgsPatternPart):
 
@@ -30,6 +61,7 @@ class PositiveNumberArgsPattern(ArgsPatternPart):
         else:
             return True
 
+
 class StringArgsPattern(ArgsPatternPart):
 
     def __init__(self, s : str):
@@ -42,6 +74,7 @@ class StringArgsPattern(ArgsPatternPart):
         if args[0] != self.s:
             raise CommandArgsPatternDoesntMatchException("%d番目の引数は「%s」である必要があります!" % (index + 1, self.s))
         return True
+
 
 class RegexArgsPattern(ArgsPatternPart):
