+++ grimbot/Commands.py
@@ -1,7 +1,52 @@
-from .CommandManager import Command
 import discord
 from typing import Sequence
 from .APIConnector import *
+from abc import abstractmethod, ABCMeta
+from .CmdPatterns import CommandLengthDoesntMatchException, ArgsPatternPart
+
+
+class Command(metaclass=ABCMeta):
+
+    def __init__(self, cmdLabel : str, argsPatternParts : Sequence[ArgsPatternPart]):
+        self.cmdLabel = cmdLabel
+        self.argsPatternParts = argsPatternParts
+
+    @abstractmethod
+    async def execute(self, args : Sequence[str], client, message : discord.Message):
+        raise NotImplementedError()
+
+    def isMatch(self, cmdStr : str) -> bool:
+        cmdSplited = cmdStr.split(" ")
+        if len(cmdSplited) == 0:
+            raise AssertionError()
+        cmdLabel = cmdSplited[0]
+        if cmdLabel != self.cmdLabel:
+            return False
+
+        args = cmdSplited[1:]
+
+        if self.requiredArgsLength() > len(args):
+            raise CommandLengthDoesntMatchException(self.requiredArgsLength())
+
+        index = 0
+        for argsPattern in self.argsPatternParts:
+            if argsPattern.argsLength != -1:
+                if argsPattern.isMatch(args[index:index + argsPattern.argsLength()], index):
+                    pass # 一致しなければ例外が発生するのでそれにまかせる
+                index += argsPattern.argsLength()
+            else:
+                if argsPattern.isMatch(args[index:]):
+                    pass # 一致しなければ例外が発生するのでそれにまかせる
+
+        return True
+
+    def requiredArgsLength(self) -> int:
+        sum = 0
+        for argsPattern in self.argsPatternParts:
+            if argsPattern.argsLength != -1:
+                sum += argsPattern.argsLength()
+        return sum
+
 
 class RainCommand(Command):
 
@@ -23,7 +68,7 @@ class RainCommand(Command):
         for address in addresses:
             APIConnector.send("", message.author.id, address, pricePerOne)
 
-        await client.send_message(message.channel, "%f Grimを%d人に、一人あたり%f Grimずつ送りました!" % (amount, len(addressDict), pricePerOne))
+        await client.send_message(message.channel, "%f Grimを%d人に、一人あたり%f Grimずつ送りました!" % (amount, len(addresses), pricePerOne))
 
     def balanceIsMoreThan(self, id : str, amount : float):
         if APIConnector.get_balance("", id) >= amount:
