+++ grimbot/__init__.py
@@ -1,3 +1,3 @@
-from .CommandManager import CommandManager, ArgsPatternPart, CommandArgsPatternDoesntMatchException, CommandLengthDoesntMatchException
 from .Commands import *
 from .CmdPatterns import *
+from .CommandManager import CommandManager
